aula07/requirements.txt
faust
kafka-python
Flask